added cache on type parsing
