+++ OrleansDashboard/DashboardMiddleware.cs
@@ -1,12 +1,12 @@
-﻿using Microsoft.AspNetCore.Http;
+﻿using System;
+using System.IO;
+using System.Reflection;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.Options;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
 using Orleans;
-using System;
-using System.IO;
-using System.Reflection;
-using System.Threading.Tasks;
 using OrleansDashboard.Client;
 using OrleansDashboard.Client.Model;
