+++ OrleansDashboard/Metrics/TypeFormatting/ParseState.cs
@@ -1,0 +1,10 @@
+﻿namespace OrleansDashboard.Metrics.TypeFormatting
+{
+    enum ParseState
+    {
+        TypeNameSection,
+        GenericCount,
+        GenericArray,
+        TypeArray
+    }
+}
