+++ OrleansDashboard/Metrics/TypeFormatting/TokenType.cs
@@ -1,0 +1,14 @@
+﻿namespace OrleansDashboard.Metrics.TypeFormatting
+{
+    enum TokenType
+    {
+        TypeNameSection,
+        GenericCount,
+        GenericArrayStart,
+        GenericArrayEnd,
+        TypeArrayStart,
+        TypeArrayEnd,
+        GenericSeparator,
+        TypeSectionSeparator
+    }
+}
