+++ PerformanceTests/DashboardGrainBenchmark.cs
@@ -1,7 +1,7 @@
-﻿using BenchmarkDotNet.Attributes;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
+using BenchmarkDotNet.Attributes;
 using OrleansDashboard.Client.Model;
 using OrleansDashboard.Metrics.History;
