+++ TestGrains/TestGrain.cs
@@ -1,6 +1,6 @@
-﻿using Orleans;
-using System;
+﻿using System;
 using System.Threading.Tasks;
+using Orleans;
 using Orleans.Runtime;
 
 namespace TestGrains
